--- Frontend/CO.CDP.OrganisationApp/Pages/Exclusions/DeclaringExclusions.cshtml
@@ -1,54 +1,0 @@
-@page "/organisation/{organisationId:guid}/forms/{formId:guid}/sections/{sectionId:guid}/declaring-exclusions"
-
-@model CO.CDP.OrganisationApp.Pages.Exclusions.DeclaringExclusionsModel
-@{
-    var hasError = ((TagBuilder)Html.ValidationMessageFor(m => m.YesNoInput)).HasInnerHtml;
-    var backUrl = $"/organisation/{Model.OrganisationId}/forms/{Model.FormId}/sections/{Model.SectionId}/add-exclusion-declaration";
-}
-<a href="@backUrl" class="govuk-back-link">Back</a>
-
-<main class="govuk-main-wrapper">
-    <div class="govuk-grid-row">
-        <div class="govuk-grid-column-two-thirds">
-            <partial name="_ErrorSummary" model="@ModelState" />
-
-            <form method="post">
-                <div class="govuk-form-group @(hasError ? "govuk-form-group--error" : "")">
-                    <fieldset class="govuk-fieldset">
-                        <legend class="govuk-fieldset__legend govuk-fieldset__legend--l">
-                            <h1 class="govuk-fieldset__heading">
-                                Do you have any exclusions to add for your organisation or a connected person?
-                            </h1>
-                        </legend>
-                        @if (hasError)
-                        {
-                            <p class="govuk-error-message">
-                                <span class="govuk-visually-hidden">Error:</span>
-                                @Html.ValidationMessageFor(m => m.YesNoInput)
-                            </p>
-                        }
-
-                        <div class="govuk-radios" data-module="govuk-radios">
-                            <div class="govuk-radios__item">
-                                <input class="govuk-radios__input" id="conditional-YES" name="YesNoInput" type="radio"
-                                       value="true" @(Model.YesNoInput == true ? "checked" : "")>
-                                <label class="govuk-label govuk-radios__label" for="conditional-YES">Yes</label>
-                            </div>
-                            <div class="govuk-radios__item">
-                                <input class="govuk-radios__input" id="conditional-NO" name="YesNoInput" type="radio"
-                                       value="false" @(Model.YesNoInput == false ? "checked" : "")>
-                                <label class="govuk-label govuk-radios__label" for="conditional-NO">No</label>
-                            </div>
-                        </div>
-
-                    </fieldset>
-                </div>
-                <div class="govuk-button-group govuk-!-margin-top-6">
-                    <button class="govuk-button" data-module="govuk-button">
-                        Continue
-                    </button>
-                </div>
-            </form>
-        </div>
-    </div>
-</main>
